refactor: use InProcessServerMethods from common-jvm
remove deps
