+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/BUILD.bazel
@@ -60,16 +60,6 @@ kt_jvm_library(
 )
 
 kt_jvm_library(
-    name = "in_process_servers_methods",
-    srcs = ["InProcessServersMethods.kt"],
-    deps = [
-        "@wfa_common_jvm//imports/java/io/grpc:api",
-        "@wfa_common_jvm//imports/java/io/grpc/inprocess",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
-    ],
-)
-
-kt_jvm_library(
     name = "event_message_flags",
     srcs = ["EventMessageFlags.kt"],
     deps = [
