+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common/server/BUILD.bazel
@@ -50,7 +50,6 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/measurement/measurementconsumer/stats:variances",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:encryption_key_pair_map",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:event_message_flags",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:in_process_servers_methods",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:kingdom_flags",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:reporting_api_server_flags",
         "//src/main/kotlin/org/wfanet/measurement/reporting/deploy/v2/common:v2alpha_flags",
